requirements.txt
telethon
openai
python-dotenv
dataclasses-json